pyproject.toml
[tool.poetry]
name = "Polyphony"
version = "2.0.0-alpha"
description = "A more robust version of PluralKit"
authors = []
license = "MIT"

[tool.poetry.dependencies]
python = "~3.9"
"discord.py" = {git = 'https://github.com/Rapptz/discord.py', rev='44abc46'}
python-dotenv = "^0.20.0"
requests = "^2.25.1"
emoji = "^2.0.0"
gitpython = "^3.1.11"
tomli = "^2.0.1"

[tool.poetry.dev-dependencies]
black = {extras = ["d"], version = "^22.6.0"}
mypy = "^0.971"

[tool.poetry.scripts]
polyphony = "polyphony:main"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
